apps/api/pyproject.toml
[tool.poetry]
name = "menuwhisper-api"
version = "0.1.0"
description = "FastAPI service for MenuWhisper"
authors = ["Oleksii Kolotenko <a1708kolot@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.34"
psycopg = {extras = ["binary"], version = "^3.2.1"}
alembic = "^1.13.2"
redis = "^5.0.7"
httpx = "^0.27.0"
tenacity = "^9.0.0"
python-multipart = "^0.0.9"
pytesseract = "^0.3.10"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
black = "^24.8.0"
pytest = "^8.3.2"
mypy = "^1.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
